% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_amendMCMC.R
\name{amendMCMC}
\alias{amendMCMC}
\title{amendMCMC}
\usage{
amendMCMC(mcmc.obj, sr.scale = 10^6)
}
\arguments{
\item{mcmc.obj}{output from a call to calcMCMCRickerBM() with output = "all"}

\item{sr.scale}{scalar used in the SR model fit (if model was fitted to Mill fish, then sr.scale = 10^6)}
}
\description{
Wrapper function that applies the calcRickerProxy() function to each MCMC sample in the calcMCMCRickerBM() output,
 calculates corresponding Sgen and fitted SR curves, and amends the output with summaries.
}
\keyword{Sgen,}
\keyword{fit}
\keyword{ricker}
