% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_VariousPlottingFunctions.R
\name{box_add}
\alias{box_add}
\alias{kernel_margin}
\title{box_add}
\usage{
box_add(perc.vec, at, width, col, bg, label = NULL, dir = "v")

kernel_margin(x, at, width, col, dir = "v")
}
\description{
These functions generate various plot additions.
}
