% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_calcDetModelFit.R
\name{calcDetModelFit}
\alias{calcDetModelFit}
\title{calcDetModelFit}
\usage{
calcDetModelFit(sr_obj, sr.scale = 10^6, min.obs = 15, resids = FALSE)
}
\arguments{
\item{sr_obj}{a data frame with Year and Spn, logRpS , and Rec (Data for 1 Stock!). Other variables can be there but are not used (RpS, Qual, ExpF etc)}

\item{min.obs}{min number of S-R pairs needed to fit a model}

\item{resids}{if TRUE, add the residuals to the output}
}
\description{
This function calculates Ricker Model parameters for spawner-recruit data using a simple linear regression of log(R/S) ~ S.  Note that these are simple deterministic model fits intended for rapid testing of input data!
}
\examples{
ricker.fit <- calcDetModelFitSR_Sample[SR_Sample$Stock == "Stock1",],min.obs = 10)
print(ricker.fit)
}
\keyword{Ricker}
\keyword{fit,}
