% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_LognormalCapacityPriorFunctions.R
\name{calcLognormalTauFromCV}
\alias{calcLognormalTauFromCV}
\alias{calcLognormalCVFromSigma}
\alias{calcLognormalTauFromSigma}
\alias{plotCapPriorCheck}
\title{calcLognormalTauFromCV}
\usage{
calcLognormalTauFromCV(cv)

calcLognormalCVFromSigma(sigma)

calcLognormalTauFromSigma(sigma)

plotCapPriorCheck(mcmc.obj, label = "Label", xlim = NULL)
}
\description{
These functions handle conversions between CV, sigma, and tau for the lognormal capacity prior.
param mcmc.obj output from calcMCMCModelFit()
}
