% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_BiolBM_Subroutines_OtherBM.R
\name{calcRickerOtherBM}
\alias{calcRickerOtherBM}
\title{calcRickerOtherBM}
\usage{
calcRickerOtherBM(X, sr.scale = 1, out.type = "Full")
}
\arguments{
\item{X}{a data frame with columns ln.a, b}

\item{sr.scale}{scalar applied to SR data in the model fitting step, need it here to scale up the Sgen values}

\item{out.type}{either "BMOnly" or "Full"}
}
\description{
This function calculates Smax and Seq for Ricker a,b parameters. Note: This function DOES NOT apply bias correction on alpha.
Whether the output is bias-corrected estimates or not depends on the par set provided by the user. This keeps the parameter
estimation and benchmark calculation steps clearly separated, given on-going debates around the bias correction.
}
\keyword{Seq,Smax}
