% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_RickerProxy.R
\name{calcRickerProxy}
\alias{calcRickerProxy}
\title{calcRickerProxy}
\usage{
calcRickerProxy(
  a,
  b,
  sd = NULL,
  spn.vals = NULL,
  sr.scale = 10^6,
  out.type = "sgen"
)
}
\arguments{
\item{a}{intercept from ln(R/S) vs S}

\item{b}{slope  from ln(R/S) vs S}

\item{spn.vals}{either NULL or a vector with spn values to use for the fitted Ricker curve. If NULL, then it generates 1000 steps between 1 fish and 1.5 *  s.eq}

\item{sr.scale}{adjust the BM outputs (e.g. if sr pars were fitted to Mill fish, then sr.scale = 10^6)}

\item{out.type}{is "sgen" for sgen only, "bm" for a list with all the BM, "curve" for just the fitted ricker curve values,}
}
\description{
This function calculates standard benchmarks and fitted SR curves from a and b values for a fit of ln(R/S) vs S (either deterministic, or one of the MCMC samples)
}
\keyword{Sgen,}
\keyword{curve}
\keyword{fitted}
