% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_RickerHelpers.R
\name{calcRickerSgen}
\alias{calcRickerSgen}
\title{calcRickerSgen}
\usage{
calcRickerSgen(a, b, sig = 1)
}
\arguments{
\item{a}{value of "a" parameter}

\item{b}{value of b parameter}
}
\description{
This function calculates Sgen a set of Ricker a,b parameters using a solver function extracted from Holt and Ogden (2013) INCLUDE LINK. This function is part of the WSPMetrics Package INCLUDE LINK, but is recreated here to avoid complications with github-based package dependencies.
}
\keyword{Sgen}
