% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_BiolBM_Subroutines_Smsy.R
\name{calcRickerSmsy}
\alias{calcRickerSmsy}
\title{calcRickerSmsy}
\usage{
calcRickerSmsy(X, method, sr.scale = 1, out.type = "Full")
}
\arguments{
\item{X}{a data frame with columns ln.a, b}

\item{method}{one of "Hilborn1985","Petermanetal2000","Scheuerell2016","BruteForce"}

\item{sr.scale}{scalar applied to SR data in the model fitting step, need it here to scale up the Sgen values}

\item{out.type}{either "SmsyOnly" or "Full"}
}
\description{
This function calculates Smsy for a Ricker a,b parameters. Note: This function DOES NOT apply bias correction on alpha.
Whether the output is bias-corrected estimates or not depends on the par set provided by the user. This keeps the parameter
estimation and benchark calculation steps clearly separated, given on-going debates around the bias correction.
}
\keyword{Smsy}
