% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_checkSRData.R
\name{checkSRData}
\alias{checkSRData}
\title{checkSRData}
\usage{
checkSRData(sr_obj, flags = NULL)
}
\arguments{
\item{sr_obj}{a data frame with  Year, Spn, Rec, and optionally SpnExp and RecAge#. (Data for 1 Stock!). Other variables can be there but are not used.}

\item{flags}{a data frame with Label, Lower, Upper. Lower and Upper define the triggers for values to be flagged. Labels have to match the criteria used in the function(LIST). The built-in object flags_default has is a template.}
}
\description{
This function calculates a set of diagnostics for the spawner-recruit data. Some apply to the whole series 
 (e.g. the contrast is spawner estimates), while others flag individual observations
 (e.g. R/S above user-specified plausible upper bound, pointing to a potential data error in either R or S).
}
\examples{
data.chk <- checkSRData(SR_Sample[SR_Sample$Stock == "Stock1",])
print(data.chk)
}
\keyword{contrast,}
\keyword{outliers}
