% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_compareBiasCorr.R
\name{compareBiasCorr}
\alias{compareBiasCorr}
\title{compareBiasCorr}
\usage{
compareBiasCorr(bm_obj, sr.scale)
}
\arguments{
\item{bm}{bm_obj output from \code{\link[RapidRicker]{calcMCMCRickerBM}}.}
}
\value{
a data frame with 1 column for each parameter (ln.a,Seq,Smsy,Sgen) and different versions in rows. Versions include: deterministic estimate with and without bias correction. Mean estimate of MCMC posteriors with (1) no bias correction, (2) bias correction using mean ln.a and mean sigma, and (3) bias correction applied to each MCMC sample and then calculate the mean. Median estimate and perentiles of MCMC posteriors with (1) no bias correction, (2)  bias correction applied to each MCMC sample.
}
\description{
This function generates a summary of raw and bias-corrected  values for SR parameters and resulting biological
benchmarks, using different approaches for implementing the bias correction.
}
\keyword{MCMC,}
\keyword{bias}
\keyword{correction,}
\keyword{posterior}
