% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_doRJAGS.R
\name{doRJAGS}
\alias{doRJAGS}
\title{doRJAGS}
\usage{
doRJAGS(
  data.obj,
  model.fn,
  settings = list(n.chains = 2, n.burnin = 20000, n.thin = 60, n.samples = 50000),
  inits,
  priors,
  pars.track,
  output = "short",
  out.path = "MCMC_Out",
  out.label = "MCMC",
  mcmc.seed = "default",
  tracing = FALSE
)
}
\arguments{
\item{data.obj}{a list object with Spn and Rec (Data for 1 Stock!), as well as the required priors (details depend on model). Other variables can be there but are not used (RpS, Qual, ExpF etc)}

\item{model.fn}{a function that defines a BUGS model}

\item{settings}{a list with n.chains (2), n.burnin (20000), n.thin (60), and n.samples (50000). Default values in brackets.}

\item{inits}{a list of lists with inits for each chain. Depens on BUGS model}

\item{priors}{a list with prior settings. De[ends on BUGS model}

\item{pars.track}{vector of text strings listing parameters to track. Depends on BUGS model}

\item{output}{one of "short" (only return summary stats for tracked parameters in a list object), "post" (also save posterior distribution samples to folder), or "all" (also produce pdf files with standard diagnostic plots)}

\item{out.path}{text string specifying  folder. if output is "post" or "all", the generated files will be stored to this folder}

\item{out.label}{label use in the output files if output is "post" or "all"}

\item{mcmc.seed}{either "default" or an integer giving the random seed for starting MCMC (R2Jags default is 123)}

\item{tracing}{if TRUE, diagnostic details for intermediate objects will be printed to the screen for debugging}
}
\description{
This function is a wrapper for calls to the jags() function of the R2jags package. It is designed for use inside the calcMCMCRickerBM() function, but can be used on its own if you are familiar with BUGS models. Check the code for calcMCMCRickerBM() to see the details. This function is adapted from forecasting code originally developed in collaboration with Sue Grant and Bronwyn MacDonald. Note: This requires installing JAGS from \href{https://sourceforge.net/projects/mcmc-jags/files/latest/download}{here}.
}
\keyword{MCMC,}
\keyword{R2Jags,}
\keyword{posterior}
