% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flags_default}
\alias{flags_default}
\title{Upper and Lower thresholds for flagging SR data issues}
\format{
A data frame with  3 required variables and optional context information (e.g. Reference):
\describe{
  \item{Label}{Labels identifyin the criteria. Have to match the criteria used in the function.}
  \item{Lower}{Any value lower than this is flagged.}
  \item{Upper}{Any value higher than this is flagged.}
  \item{}{}
  ...
}
}
\source{
Include REF
}
\usage{
flags_default
}
\description{
A dataset containing observed recruits/spawner and various environmental covariates used in the 2019 Forecast for 
Fraser River Sockeye Salmon (REF).IMPORTANT NOTE: Time series are not lined up based on plausible mechanism (e.g. sea surface temp may affect overall productivity during early ocean entry, but RpEff values are by brood year. Stock-specific offsets are required to align values properly. INCLUDE LIST!
}
\keyword{datasets}
