% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_generateInits.R
\name{generateInits}
\alias{generateInits}
\title{generateInits}
\usage{
generateInits(priors, model_type = "Basic", n.chains = 2, filename = NULL)
}
\arguments{
\item{priors}{output from \code{\link[RapidRicker]{generatePriors}}}

\item{model_type}{one of "Basic", "Kalman", or "AR1".  For details, see \href{https://github.com/SOLV-Code/RapidRicker/wiki/3:--Ricker-Model-Forms-in-BUGS-JAGS}{this wiki}.}

\item{n.chains}{number of MCMC chains to initiate}

\item{filename}{either NULL, or a file/path for an output file to save the generated priors (e.g. "OUTPUT/Inits.txt")}

\item{sr.scale}{an integer value used to rescale the Spn and Rec variables in sr_obj, prior to the MCMC fit, default = 10^6 (i.e. convert to millions). NOTE: If sr.scale is different from 1, then
the benchmark estimates are scaled back, but the MCMC estimates of alpha and beta will be in different units then the alpha and beta estimates from the deterministic fit.}
}
\description{
This function generates  initial values ("inits") for the MCMC sampling forms. Defaults settings are under development. Follow the discussion at \href{https://github.com/SOLV-Code/RapidRicker/issues/71}{this thread}.
}
\keyword{inits}
