% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gls_test_data}
\alias{gls_test_data}
\title{Spawner Recruit Test Data}
\format{
A data frame with  5 required variables:
\describe{
  \item{Stock}{ stock name}
  \item{Year}{brood year}
  \item{Spn}{spawner estimate}
  \item{Rec}{recruitment estimate}
  \item{logRpS}{calculated value of log(Rec/Spn), usually with mutate(logRpS = log(Rec/Spn))}
  \item{}{}
  ...
}
}
\source{
Dummy Data
}
\usage{
gls_test_data
}
\description{
A dataset containing spawners and  recruits by brood year for 2 stocks.
}
\keyword{datasets}
