% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_PlotRanking.R
\name{plotRanking}
\alias{plotRanking}
\title{plotRanking}
\usage{
plotRanking(
  data.df,
  trim = 25,
  maxvars = NULL,
  xlim = NULL,
  flag = NULL,
  mean.pt = FALSE
)
}
\arguments{
\item{data.df}{a data frame where each column is a set of values. Column labels become plot labels}

\item{trim}{an integer value between 0 and 49 to specify the percentile to be plotted for the whiskers (e.g. trim = 25 plots upper/lower quartiles).}

\item{maxvars}{integer specifying the maximum number of items  to plot (e.g. 30 = plot top 30)}

\item{xlim}{vector to specify the range for the horizontal axis}

\item{flag}{optional text string specifying a column to highlight inthe ranked plot (e.g. a Stock Name)}

\item{mean.pt}{if TRUE, show mean instead of median.}
}
\description{
This function generates a plot showing range and medians for columns in a table, ranked by median or mean.  Original function created for the \href{https://github.com/SOLV-Code/SPATFunctions-Package}{SPATFunctions Package}, but adapted here to customize formatting for report figures and avoid adding a dependency.
}
\keyword{plot}
\keyword{rank,}
