% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_RickerProxy.R
\name{plotRickerProxy}
\alias{plotRickerProxy}
\title{plotRickerProxy}
\usage{
plotRickerProxy(proxy.obj, x.lim = NULL, axis.scale = 1)
}
\arguments{
\item{proxy.obj}{output from  call to calcRickerProxy with out = "curve" or "all"}

\item{x.lim}{vector with lowerr and upper bound for the plot}

\item{axis.scale}{rescales the axes}
}
\description{
This function plots fitted SR curve and standard benchmarks
}
\keyword{Ricker}
\keyword{benchmarks}
\keyword{curve,}
