% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_PlotTornado.R
\name{plotTornado}
\alias{plotTornado}
\title{plotTornado}
\usage{
plotTornado(
  data.df,
  keep.rank = FALSE,
  add.labels = TRUE,
  xlim = NULL,
  solid.refline = NULL,
  dashed.refline = NULL
)
}
\arguments{
\item{data.df}{a data frame with columns Label, Lower, Mid, Upper}

\item{add.labels}{if = TRUE, data.df needs another column called AddLabel which is added to the plot a labels inside the label axis (e.g. Avg abd as context)}

\item{rank}{default rank is based on range between Upper and Lower. if keep.rank = TRUE, then keep the input order as is.}
}
\description{
This function generates a plot showing a range for each row. Typically each row is a stock, the first col is a stock label, and the remaining columns have values for lower, mid, and upper (e.g. from a sensitivity test)
}
\keyword{plot}
\keyword{sensitivity}
\keyword{test,}
