% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODEL_ricker_AR1_BUGS.R
\name{ricker.ar1.BUGS}
\alias{ricker.ar1.BUGS}
\title{ricker.BUGS}
\usage{
ricker.ar1.BUGS()
}
\description{
This function is the BUGS version of the Ricker model fit and associated benchmarks. It is used inside
of calcMCMCRickerBM() which passes it to  doRJAGS(), which is a wrapper for the jags() function from the R2jags package. For details such as model form and variable definitions, refer to \href{https://github.com/SOLV-Code/RapidRicker/wiki/MCMC-Using-R2Jags}{this wiki page}. The code is modified from code by Catherine Michielsens and expanded for AR1 based on Eq21 and 22 of Fleischman and Evenson (2010; ADFG FMS10-04) .
An example of previous use is \href{https://www.dfo-mpo.gc.ca/csas-sccs/publications/resdocs-docrech/2015/2015_048-eng.html}{Taku Coho}.
This function has no arguments, because the jags call sets up the inputs.
}
