% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_RickerHelpers.R
\name{ricker.array}
\alias{ricker.array}
\title{ricker.array}
\usage{
ricker.array(a.vec, b.vec, spn.vec)
}
\arguments{
\item{a.vec}{vector of "a" parameters, typically an MCMC posterior sample}

\item{b.vec}{vector of "b" parameters, typically an MCMC posterior sample
#' @param spn.vec vector of spawner abundances for which to calculate a distribution of recruits, based on the a and b samples}
}
\description{
This function calculates Recruits for a set of Ricker a,b parameters and spawner abundance (by setting up 2 dim arrays).
}
\keyword{Ricker}
\keyword{curve}
