% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODEL_ricker_TimeVaryingAlpha_BUGS.R
\name{ricker.varalpha.BUGS}
\alias{ricker.varalpha.BUGS}
\title{ricker.varalpha.BUGS}
\usage{
ricker.varalpha.BUGS()
}
\description{
This function is the BUGS version of the Ricker model fit with time-varying productivity (a.k.a Kalman Filter, Recursive Bayes) and associated benchmarks. It is used inside 
of calcMCMCRickerBM() which passes it to  doRJAGS(), which is a wrapper for the jags() function from the R2jags package. For details such as model form and variable definitions, refer to \href{https://github.com/SOLV-Code/RapidRicker/wiki/MCMC-Using-R2Jags}{this wiki page}.
This function has no arguments, because the jags call sets up the inputs.  The code is modified from code by Catherine Michielsens. An example of previous use is \href{https://www.dfo-mpo.gc.ca/csas-sccs/publications/resdocs-docrech/2015/2015_048-eng.html}{Taku Coho}.
}
