% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SR_Sample}
\alias{SR_Sample}
\title{Sample Spawner Recruit Data}
\format{

}
\source{
Dummy Data
}
\usage{
SR_Sample
}
\description{
A dataset containing spawners, recruits, spawner estimate expansions, and age comp by brood year for mutiple stocks.
}
\keyword{datasets}
