% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_VariousPlottingFunctions.R
\name{box_add}
\alias{box_add}
\title{Plotting Functions}
\usage{
box_add(perc.vec, at, width, col, bg, label)
}
\description{
These functions create various plot components.
}
\keyword{plots}
