% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_calcMCMCRickerBM.R
\name{calcMCMCRickerBM}
\alias{calcMCMCRickerBM}
\title{calcMCMCRickerBM}
\usage{
calcMCMCRickerBM(
  fit_obj,
  sr.scale = 10^6,
  Smsy.method = "Scheuerell2016",
  Sgen.method = "Connorsetal2022",
  drop.resids = TRUE
)
}
\arguments{
\item{fit_obj}{a list object are not used (RpS, Qual, ExpF etc)}

\item{sr.scale}{one of "HoltOgden2013", "samSim", "Connorsetal2022","BruteForce". Default is "Connorsetal2022"}

\item{Smsy.method}{one of  "Hilborn1985","Petermanetal2000","Scheuerell2016","BruteForce". Default is "Scheuerell2016".}

\item{drop.resids}{if TRUE, exclude the annual residuals from the output}
}
\description{
This function calculates calculates standard biological benchmarks (Smsy, Seq, Smax, Umsy),
applying the subroutines \code{\link[RapidRicker]{calcRickerSmsy}}, \code{\link[RapidRicker]{calcRickerSgen}},
and \code{\link[RapidRicker]{calcRickerOtherBM}} with user-specified settings.
}
\examples{
ricker.bm <- calcMCMCRickerBM(SR_Sample[SR_Sample$Stock == "Stock1",],min.obs = 10)
print(ricker.bm)
}
\keyword{Bayesian,}
\keyword{MCMC,}
\keyword{Ricker}
\keyword{Seq,}
\keyword{Smax,}
\keyword{Smsy,}
\keyword{Umsy}
\keyword{fit,}
\keyword{posterior,}
