% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_BiolBM_Subroutines_Sgen.R
\name{calcRickerSgen}
\alias{calcRickerSgen}
\title{calcRickerSgen}
\usage{
calcRickerSgen(X, method = "Connorsetal2022", sr.scale = 1, out.type = "Full")
}
\arguments{
\item{X}{a data frame with columns ln.a, b, sigma, and optionally Smsy}

\item{method}{one of "HoltOgden2013", "samSim", "Connorsetal2022","BruteForce"}

\item{sr.scale}{scalar applied to SR data in the model fitting step, need it here to scale up the Sgen values}

\item{out.type}{either "BMOnly" or "Full"}
}
\description{
This function calculates Sgen for a set of Ricker ln.a,b,sigma parameters, and optionally Smsy.
NOTE: If method is "HoltOgden2013", then Smsy is always calculated based on Hilborn (1985) approximation,
and if Smsy is provided, it will give a warning that it was ignored. Note: This function DOES NOT apply bias correction on alpha.
Whether the output is bias-corrected estimates or not depends on the par set provided by the user. This keeps the parameter
estimation and benchark calculation steps clearly separated, given on-going debates around the bias correction.
}
\keyword{Sgen}
