% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flags_default}
\alias{flags_default}
\title{Upper and Lower thresholds for flagging SR data issues}
\format{
A data frame with  3 required variables and optional context information (e.g. Reference):
\describe{
  \item{Label}{Labels identifyin the criteria. Have to match the criteria used in the function.}
  \item{Lower}{Any value lower than this is flagged.}
  \item{Upper}{Any value higher than this is flagged.}
  \item{}{}
  ...
}
}
\source{
Include REF
}
\usage{
flags_default
}
\description{
A dataset containing labels, thresholds and decriptions for various data qulity metrics. INCLUDE LIST!
}
\keyword{datasets}
