% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_RapidRicker.R
\name{RapidRicker}
\alias{RapidRicker}
\title{RapidRicker}
\usage{
RapidRicker(sr_obj_m, min.obs = 10, trace = TRUE)
}
\arguments{
\item{sr_obj_m}{a data frame with Stock, Year and Spn, logRpS (Data for multiple stocks!). Other variables can be there but are not used (RpS, Qual, ExpF etc).}

\item{min.obs}{min number of S-R pairs needed to fit a model}

\item{trace}{if TRUE, print various intermediate diagnostic output to the console}
}
\description{
This function is a wrapper for testDetRickerBM(). It loops throug multiple stocks and applies all the sensitivity test variations, then calculates the perc differences to base case. For calculation details, refer to calcDetRickerBM and testDetRickerBM.
}
\examples{
rapid.ricker.out <- RapidRicker(SR_Sample,min.obs = 10,  trace=FALSE)
}
\keyword{sensitivity}
\keyword{test}
