% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_testDetRickerBM.R
\name{testDetRickerBM}
\alias{testDetRickerBM}
\title{testDetRickerBM}
\usage{
testDetRickerBM(sr_obj, min.obs = 15, type = "jack", trace = FALSE)
}
\arguments{
\item{sr_obj}{a data frame with Year and Spn, logRpS (Data for 1 Stock!). Other variables can be there but are not used (RpS, Qual, ExpF etc).}

\item{min.obs}{min number of S-R pairs needed to fit a model}

\item{type}{one of "jack" (for a drop 1 jackknife test), "retro" (retrospective test starting with min.obs, then adding more years), or "revretro" (reverse retrospective, starting with all years and them dropping the earlier obs one at a time until only the most recent min.obs are left).}

\item{trace}{if TRUE, print various intermediate diagnostic output to the console}
}
\description{
This function calculates standard biological benchmarks (Smsy, Seq, Smax, Umsy) for different subsets of the input data, using the function calcDetRickerBM(). See calculation details there.
}
\examples{
ricker.test <- testDetRickerBM(SR_Sample[SR_Sample$Stock == "Stock1",],min.obs = 10,  type="retro")
}
\keyword{sensitivity}
\keyword{test}
